+++ web/modals.js
@@ -2,24 +2,23 @@ import ConfirmationModal from "./common/Confirmation";
 import TeamOrSoloChoiceModal from "./pwa/components/TeamOrSoloChoice";
 import NewTeamMateModal from "./pwa/components/NewTeamMate";
 import EndMissionModal from "./pwa/components/EndMission";
-import ExpenditureDialogModal from "./pwa/components/ExpenditureDialog";
 import ApiErrorDialogModal from "./pwa/components/ApiErrorDialog";
 import CGUModal from "./landing/cgu";
 import UserReadQRCodeModal from "./control/UserReadQRCode";
 import UnavailableOfflineModeModal from "./common/UnavailableOfflineMode";
 import KilometerReadingModal from "./pwa/components/KilometerReadingModal";
-import PDFExport from "./pwa/components/PDFExport";
-import WarningEndMissionModal from "./pwa/components/WarningEndMissionModal/WarningEndMissionModal";
 import NewsletterSubscriptionModal from "./landing/NewsletterSubscription";
 import GeolocPermissionInfoModal from "./pwa/components/GeolocPermissionInfoModal";
 import TypeformModal from "./pwa/components/TypeformModal";
 import LogHolidayModal from "./pwa/components/LogHoliday";
 
+// controller
 import ControllerHelp from "./controller/components/modals/ControllerHelp";
 import ControllerExportC1BAll from "./controller/components/modals/ControllerExportC1BAll";
 import ControllerExportC1BOne from "./controller/components/modals/ControllerExportC1BOne";
 import ConfirmationCancelControlBulletinModal from "./controller/components/modals/ConfirmationCancelControlBulletinModal";
 
+// admin
 import UpdateCompanyDetailsModal from "./admin/modals/UpdateCompanyDetailsModal";
 import C1BExportModal from "./admin/modals/C1BExportModal";
 import ExcelExportModal from "./admin/modals/ExcelExportModal";
@@ -29,11 +28,13 @@ import SelectEmployeeModal from "./admin/modals/SelectEmployeeModal";
 import TerminateEmploymentModal from "./admin/modals/TerminateEmploymentModal";
 import BatchInviteModal from "./admin/modals/BatchInviteModal";
 
+// home
 import ChangeEmailModal from "./pwa/modals/ChangeEmailModal";
 import ChangeNameModal from "./pwa/modals/ChangeNameModal";
 import ChangePhoneNumberModal from "./pwa/modals/ChangePhoneNumberModal";
 import ChangeTimezoneModal from "./pwa/modals/ChangeTimezoneModal";
 
+// employee
 import FirstActivitySelectionModal from "./pwa/modals/FirstActivitySelectionModal";
 import TeamSelectionModal from "./pwa/modals/TeamSelectionModal";
 import NewMissionModal from "./pwa/modals/NewMissionModal";
@@ -41,6 +42,9 @@ import DriverSelectionModal from "./pwa/modals/DriverSelectionModal";
 import CommentInputModal from "./pwa/modals/CommentInputModal";
 import ActivityRevisionOrCreationModal from "./pwa/modals/ActivityRevision";
 import UpdateVehicleModal from "./pwa/modals/VehicleBookingModal";
+import WarningEndMissionModal from "./pwa/modals/WarningEndMissionModal";
+import PDFExportModal from "./pwa/modals/PDFExportModal";
+import ExpenditureModal from "./pwa/modals/ExpenditureModal";
 
 export const MODAL_DICT = {
   firstActivity: FirstActivitySelectionModal,
@@ -55,11 +59,11 @@ export const MODAL_DICT = {
   updateVehicle: UpdateVehicleModal,
   dataExport: ExcelExportModal,
   tachographExport: C1BExportModal,
-  pdfExport: PDFExport,
+  pdfExport: PDFExportModal,
   teamOrSoloChoice: TeamOrSoloChoiceModal,
   newTeamMate: NewTeamMateModal,
   endMission: EndMissionModal,
-  expenditures: ExpenditureDialogModal,
+  expenditures: ExpenditureModal,
   apiErrorDialog: ApiErrorDialogModal,
   cgu: CGUModal,
   changeEmail: ChangeEmailModal,
